requirements.txt
pandas
numpy
requests
beautifulsoup4
yfinance
matplotlib
seaborn
adjustText
python-dotenv
openpyxl
lxml